+++ src/ONTraC/bin/ONTraC_analysis.py
@@ -16,7 +16,7 @@ from ..version import __version__
 # ------------------------------------
 # Constants
 # ------------------------------------
-IO_OPTIONS = ['dataset', 'preprocessing_dir', 'GNN_dir', 'NTScore_dir']
+IO_OPTIONS = ['input', 'preprocessing_dir', 'GNN_dir', 'NTScore_dir']
 
 
 # ------------------------------------
@@ -72,7 +72,10 @@ def add_embedding_adjust_group(optparser: OptionParser) -> None:
         '--sigma',
         dest='sigma',
         type='float',
-        help='Sigma for the exponential function. Default is the mean of the distances between the cell type pairs.')
+        default=1,
+        help=
+        'Sigma for the exponential function to control the similarity between different cell types or clusters. Default is 1.'
+    )
     optparser.add_option_group(group)