+++ src/ONTraC/niche_net/__init__.py
@@ -1,1 +1,2 @@
-from ._niche_net import construct_niche_network, gen_samples_yaml, ct_coding_adjust
+from ._niche_net import construct_niche_network, ct_coding_adjust, gen_samples_yaml, get_embedding_columns
+from .preprocessing import gen_original_data