+++ src/ONTraC/optparser/_IO.py
@@ -16,8 +16,29 @@ def add_IO_options_group(optparser: OptionParser, io_options: Optional[List[str]
         return
     # I/O options group
     group_io = OptionGroup(optparser, "IO")
-    if 'dataset' in io_options:
-        group_io.add_option('-d', '--dataset', dest='dataset', type='string', help='Original input dataset.')
+    if 'input' in io_options:
+        group_io.add_option(
+            '--meta-input',
+            dest='meta_input',
+            type='string',
+            help=
+            'Meta data file in csv format. Each row is a cell and each column is a meta data. The first column should be the cell name. Coordinates (x, y) and sample should be included. Cell type is optional.'
+        )
+        group_io.add_option(
+            '--exp-input',
+            dest='exp_input',
+            type='string',
+            help=
+            'Normalized expression file in csv format. Each row is a cell and each column is a gene. The first column should be the cell name. If not provided, cell type should be included in the meta data file.'
+        )
+        group_io.add_option('--embedding-input',
+                            dest='embedding_input',
+                            type='string',
+                            help='Embedding file in csv format. The first column should be the cell name.')
+        group_io.add_option('--decomposition-input',
+                            dest='decomposition_input',
+                            type='string',
+                            help='Decomposition file in csv format. The first column should be the cell name.')
     if 'preprocessing_dir' in io_options:
         group_io.add_option('--preprocessing-dir',
                             dest='preprocessing_dir',
@@ -26,7 +47,10 @@ def add_IO_options_group(optparser: OptionParser, io_options: Optional[List[str]
     if 'GNN_dir' in io_options:
         group_io.add_option('--GNN-dir', dest='GNN_dir', type='string', help='Directory for the GNN output.')
     if 'NTScore_dir' in io_options:
-        group_io.add_option('--NTScore-dir', dest='NTScore_dir', type='string', help='Directory for the NTScore output.')
+        group_io.add_option('--NTScore-dir',
+                            dest='NTScore_dir',
+                            type='string',
+                            help='Directory for the NTScore output.')
 
     optparser.add_option_group(group_io)
 
@@ -36,24 +60,58 @@ def validate_io_options(optparser: OptionParser,
                         io_options: Optional[List[str]],
                         overwrite_validation: bool = True) -> None:
     """Validate IO options from a OptParser object.
-
+    :param optparser: OptionParser object.
+    :param options: Values object.
+    :param io_options: List of IO options.
+    :param overwrite_validation: Overwrite validation.
     Ret: None
     """
     if io_options is None:
         return
-    if 'dataset' in io_options:
-        if not options.dataset:
-            error('Please provide a dataset.')
+    if 'input' in io_options:
+        # meta data
+        if not options.meta_input:
+            error('Please provide a meta data file in csv format.')
             optparser.print_help()
             sys.exit(1)
-        if not os.path.isfile(options.dataset):
-            error(f'The input file ({options.dataset}) you given does not exist.')
+        if not os.path.isfile(options.meta_input):
+            error(f'The meta data file ({options.meta_input}) you given does not exist.')
             optparser.print_help()
             sys.exit(1)
-        if not options.dataset.endswith(('csv', 'csv.gz')):
-            error(f'The input file ({options.dataset}) should be in csv format.')
+        if not options.meta_input.endswith(('csv', 'csv.gz')):
+            error(f'The meta data file ({options.meta_input}) should be in csv format.')
             optparser.print_help()
             sys.exit(1)
+        # expression data
+        if options.exp_input:
+            if not os.path.isfile(options.exp_input):
+                error(f'The expression data file ({options.exp_input}) you given does not exist.')
+                optparser.print_help()
+                sys.exit(1)
+            if not options.exp_input.endswith(('csv', 'csv.gz')):
+                error(f'The expression data file ({options.exp_input}) should be in csv format.')
+                optparser.print_help()
+                sys.exit(1)
+        # embedding
+        if options.embedding_input:
+            if not os.path.isfile(options.embedding_input):
+                error(f'The embedding file ({options.embedding_input}) you given does not exist.')
+                optparser.print_help()
+                sys.exit(1)
+            if not options.embedding_input.endswith(('csv', 'csv.gz')):
+                error(f'The embedding file ({options.embedding_input}) should be in csv format.')
+                optparser.print_help()
+                sys.exit(1)
+        # decomposition
+        if options.decomposition_input:
+            if not os.path.isfile(options.decomposition_input):
+                error(f'The decomposition file ({options.decomposition_input}) you given does not exist.')
+                optparser.print_help()
+                sys.exit(1)
+            if not options.decomposition_input.endswith(('csv', 'csv.gz')):
+                error(f'The decomposition file ({options.decomposition_input}) should be in csv format.')
+                optparser.print_help()
+                sys.exit(1)
 
     if 'preprocessing_dir' in io_options:
         if not options.preprocessing_dir:
@@ -107,11 +165,17 @@ def write_io_options_memo(options: Values, io_options: Optional[List[str]]) -> N
     if io_options is None:
         return
     info('            -------- I/O options -------             ')
+    if 'input' in io_options:
+        info(f'meta data file:  {options.meta_input}')
+        if options.exp_input:
+            info(f'expression data file:  {options.exp_input}')
+        if options.embedding_input:
+            info(f'embedding file:  {options.embedding_input}')
+        if options.decomposition_input:
+            info(f'decomposition file:  {options.decomposition_input}')
     if 'preprocessing_dir' in io_options:
         info(f'preprocessing output directory:  {options.preprocessing_dir}')
     if 'GNN_dir' in io_options:
         info(f'GNN output directory:  {options.GNN_dir}')
     if 'NTScore_dir' in io_options:
         info(f'NTScore output directory:  {options.NTScore_dir}')
-    if 'dataset' in io_options:
-        info(f'dataset: {options.dataset}')